+++ src/GafferUIModule/ImageGadgetBinding.cpp
@@ -52,7 +52,5 @@ void GafferUIModule::bindImageGadget()
 	GadgetClass<ImageGadget>()
 		.def( init<const std::string &>() )
 		.def( init<IECoreImage::ConstImagePrimitivePtr>() )
-		.def( "textureLoader", &ImageGadget::textureLoader, return_value_policy<CastToIntrusivePtr>() )
-		.staticmethod( "textureLoader" )
 	;
 }